upload: Fix the call to alert_message with the correct id
alert_message was updated to include the div id from the
HTML element, and without this, the alert_message function
cannot be called and the program just stops there.
This fix makes it so that the alert message is called properly
and the file browser input field is reset as expected.